feat(Status): Refactor service details data handling into a custom hook for better code reuse
